Added possibility to manage group members of a group
We're using onyxpoint/gpasswd to enable the management of group members
(for example ad users as members of a local group). The
onyxpoint/gpasswd module does that by overwriting parts of the group
provider. Changes to user.pp was necessary so that members of the
primary group of a user could be managed as well.