Add a source to the home directories.
This allows admins to put certain files for certain users in a directory and have them managed Puppet. For this to work, you'd need to make sure  exists, and that  is set to 'remote' for the users that need their home directory contents managed but not purged (true, if you also want to purge unmanaged files).