Move `salts` fact logic into the `setcode` block
Windows machines are having a hard time if this module is in use on
a mixed Windows/Linux environment since they are still getting this
module's custom facts. The `salts` fact tries to read the shadow file
but that file is not there on Windows, and then it tries to split that
nil value as you would with a string.

This commit moves that shadow file handling into the `setcode` block
which should only be executed in case the confinement passes - which
it should not on Windows or any other machines where the shadow file
does not exist for any reason.

Fixes deric/puppet-accounts#75.
Expands on the comment on deric/puppet-accounts#76.