Properly Lint `metadata.json`
We are having test failures based on the official `master` branch
failing. One of those is caused by the `metadata.json` containing
a reference to the Puppet Enterprise version requirements.

That requirement is now being ignored by the Forge and it instead
uses the base Puppet version and matches that up with the Pupppet
Enterprise version. Long story short, that is being done since the
Puppet Enterprise versions do not follow semantic versioning anymore.
See the link below for the official announcement.

https://groups.google.com/d/msg/puppet-users/nkRPvG4q0Oo/GmXa109aJQAJ